Move the subfiles into the src folder and fix inconsitent imports
