+++ lib/src/component_animated_builder.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:mvvm_builder/presenter_builder.dart';
 
 import 'component_builder.dart';
 import 'mvvm_context.dart';
 import 'mvvm_model.dart';
+import 'presenter_builder.dart';
 
 /// -----------------------------------------------
 /// SINGLE ANIMATION CONTENT WIDGET
