+++ lib/src/mvvm_context.dart
