+++ lib/src/mvvm_model.dart
